Move UnreadWidgetProvider back to old location
