+++ app/k9mail/src/main/java/com/fsck/k9/provider/UnreadWidgetProvider.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.widget.unread
+package com.fsck.k9.provider
 
 import android.app.PendingIntent
 import android.appwidget.AppWidgetManager
@@ -10,8 +10,19 @@ import android.widget.RemoteViews
 import com.fsck.k9.EarlyInit
 import com.fsck.k9.R
 import com.fsck.k9.inject
+import com.fsck.k9.widget.unread.UnreadWidgetConfigurationActivity
+import com.fsck.k9.widget.unread.UnreadWidgetData
+import com.fsck.k9.widget.unread.UnreadWidgetRepository
 import timber.log.Timber
 
+/**
+ * Unread home screen widget "provider"
+ *
+ * IMPORTANT: This class must not be renamed or moved, otherwise unread widgets added to the home screen using an older
+ * version of the app will stop working.
+ *
+ * The rest of the unread widget specific code can be found in the package [com.fsck.k9.widget.unread].
+ */
 class UnreadWidgetProvider : AppWidgetProvider(), EarlyInit {
     private val repository: UnreadWidgetRepository by inject()
