+++ app/k9mail/src/main/java/com/fsck/k9/widget/unread/UnreadWidgetUpdater.kt
@@ -4,6 +4,7 @@ import android.appwidget.AppWidgetManager
 import android.content.ComponentName
 import android.content.Context
 import android.content.Intent
+import com.fsck.k9.provider.UnreadWidgetProvider
 
 class UnreadWidgetUpdater(private val context: Context) {
     private val appWidgetManager = AppWidgetManager.getInstance(context)
